pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Just another agent"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "sqlite-vec>=0.1.6",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.30.0",
]
all = [
    "openai>=1.0.0",
    "anthropic>=0.30.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
